Time Series Data Library metadata.

The metadata the Time Series Data Library.

meta_tsdl

Format

meta_tsdl is a data frame containing information about the tsdl series with the following structure:

source

Source of the time series.

description

A short description of the time series.

frequency

Frequency of the time series.

start

Start year of the time series.

subject

Subject (or type) of the time series.

References

Hyndman, R.J. "Time Series Data Library", https://datamarket.com/data/list/?q=provider:tsdl.

See also

Examples

str(meta_tsdl)

#> 'data.frame':	648 obs. of  5 variables:
#>  $ source     : chr  "Abraham & Ledolter (1983)" "Abraham & Ledolter (1983)" "Abraham & Ledolter (1983)" "Abraham & Ledolter (1983)" ...
#>  $ description:List of 648
#>   ..$ : chr "Quarterly Iowa nonfarm income (1948 – 1979)"
#>   ..$ : chr "Monthly demand repair parts large/heavy equip. Iowa 1972 – 1979"
#>   ..$ : chr "Montly av. residential gas usage Iowa (cubic feet)*100 ’71 – ’79"
#>   ..$ : chr "Monthly gasoline demand Ontario gallon millions 1960 – 1975"
#>   ..$ : chr "Monthly % changes in Canadian wages and salaries ’67 – ’75"
#>   ..$ : chr "Monthly sales of U.S. houses (thousands) 1965 – 1975"
#>   ..$ : chr "Quarterly growth rates of Iowa nonfarm income (1948 – 1979)"
#>   ..$ : chr "Monthly Av. residential electricity usage Iowa city 1971 – 1979"
#>   ..$ : chr "Monthly car sales in Quebec 1960-1968"
#>   ..$ : chr "Monthly traffic fatalities in Ontario 1960-1974"
#>   ..$ : chr "Monthly U.S. housing starts (privately owned 1-family) 1965 – 1975"
#>   ..$ : chr "Quarterly U.S. new plant/equip. expenditures ’64 – ’76 billions"
#>   ..$ : chr "Four-weekly totals of beer shipments"
#>   ..$ : chr "Monthly diffs yields: mortgages and govt. loans, Holland ’61 – ’74"
#>   ..$ : chr "More advertising and sales data: 36 consecutive monthly sales and advertising expenditures of a dietary weight control product"
#>   ..$ : chr "Monthly U.S. male (16-19 years) unemployment figures (thousands) 1948-1981"
#>   ..$ : chr "Monthly U.S. male (20 years and over) unemployment figures (10**3) 1948-1981"
#>   ..$ : chr "Monthly U.S. female (16-19 years) unemployment figures (thousands) 1948-1981"
#>   ..$ : chr "Monthly U.S. female (20 years and over) unemployment figures (10**3) 1948-1981"
#>   ..$ : chr "Zuerich monthly sunspot numbers 1749-1983"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 2B"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 3"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 5"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 6"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 7"
#>   ..$ : chr "Annual number of lynx pelts (Hudson’s Bay company, Canada) 1857 – 1911"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 8"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 9"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 10"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 11"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 12"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 13"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 14"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 15"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 16"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 17"
#>   ..$ : chr "Annual unit price of lynx pelts (Hudson’s Bay company, Canada) 1857 – 1911"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 18"
#>   ..$ : chr "Annual yield of grain on Broadbalk field at Rothamsted 1852-1925: plot 19"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 2B"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 3"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 5"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 6"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 7"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 8"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 9"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 10"
#>   ..$ : chr "Monthly mean thickness (Dobson units) ozone column Arosa, Switzerland 1926-1971"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 11"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 12"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 13"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 14"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 15"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 16"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 17"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 18"
#>   ..$ : chr "Annual yield of straw on Broadbalk field at Rothamsted 1852-1925: plot 19"
#>   ..$ : chr "Annual changes in the earth’s rotation, day length (sec*10**-5) 1821-1970"
#>   ..$ : chr "Annual male melanoma incidence (age-adjusted per 10**5) Connecticut 1936-1972"
#>   ..$ : chr "Annual total melanoma incidence (age-adjusted per 10**5) Connecticut 1936-1972"
#>   ..$ : chr "Annual sunspot relative number 1936-1972"
#>   ..$ : chr "Monthly Canadian total unemployment figures (thousands) 1956-1975"
#>   ..$ : chr "Half-hourly precipitation and stream flow, River Hirnant, Wales, UK, November and December 1972."
#>   ..$ : chr "Sacramento River at Keswick, California, Oct. 1939 – Sep. 1960"
#>   ..$ : chr "Clearwater River at Kamiah, Idaho. 1911 – 1965"
#>   ..$ : chr "Judith River near Utica, MT. 1920 – 1960"
#>   ..$ : chr "Madison River near west Yellowstone, MT. 1923 – 1960"
#>   ..$ : chr "Whiterocks River near Whiterocks, Utah, 1930 – 1960"
#>   ..$ : chr "Middle Boulder Creek at Nederland, CO. 1912 – 1960"
#>   ..$ : chr "South Platte River below Cheesman Lake, CO. 1925 – 1960"
#>   ..$ : chr "Neches River near Rockland, Texas. 1940 – 1960"
#>   ..$ : chr "Big Ford River at big falls, MN. 1929 – 1960"
#>   ..$ : chr "Skunk River at Augusta, Iowa. 1915 – Sep. 1960"
#>   ..$ : chr "Current River at Van Buren, MO. 1922 – Sep. 1960"
#>   ..$ : chr "Trinity River at Lewiston, California, Oct. 1912 – Sep. 1960"
#>   ..$ : chr "Wolf River at New London, WI. 1914 – 1960"
#>   ..$ : chr "Mad River near Springfield, OH. 1915 – Sep. 1960"
#>   ..$ : chr "West Branch Delaware River at Hale Eddy, NY. 1916 – Sep. 1960"
#>   ..$ : chr "Pemigewasset River at Plymouth, NH. 1904 – 1960"
#>   ..$ : chr "Rappahannock River near Fredericksburg, VA. 1911 – 1960"
#>   ..$ : chr "James River at Buchanan, VA. 1911 – 1960"
#>   ..$ : chr "Oostanaula River at Resaca, GA. 1893 – 1960"
#>   ..$ : chr "Feather River at Oroville, California, Oct. 1902 – Sep. 1977"
#>   ..$ : chr "American River at Fair Oaks, California, Oct. 1906 – Sep. 1960"
#>   ..$ : chr "Eel River above Dos Rios, California, Oct. 1952 – Sep. 1960"
#>   ..$ : chr "Rock Creek at Little Round Valley, NR. Bishop, California, Sep. 1960"
#>   ..$ : chr "McKenzie River at McKenzie Bridge, Oregon, Oct. 1911 – Sep. 1960"
#>   ..$ : chr "S. F. Skykomish River near Index, Washington, Oct. 1923 – Sep. 1960"
#>   ..$ : chr "Boise River near Twin Springs, Idaho, Oct. 1912 – Sep. 1960"
#>   ..$ : chr "Mean maximum temperature in Melbourne: degrees C. Jan 71 – Dec 90."
#>   ..$ : chr "Daily maximum temperatures in Melbourne, Australia, 1981-1990"
#>   ..$ : chr "Daily minimum temperatures in Melbourne, Australia, 1981-1990"
#>   ..$ : chr "Daily rainfall in Melbourne, Australia, 1981-1990"
#>   ..$ : chr "Col 1: Quarterly average weekly male earnings in Australia, all industries. Col 2: CPI for same quarters. Col 3"| __truncated__
#>   ..$ : chr "Total building and construction activity in Australia: approvals each quarter in $m at 1989/90 prices. Sep 1973 – Mar 1995"
#>   ..$ : chr "Monthly basic iron production in Australia: thousand tonnes. Jan 1956 – Aug 1995"
#>   ..$ : chr "Basic quarterly iron production in Australia: thousand tonnes. Mar 1956 – Sep 1994"
#>   ..$ : chr "Monthly beer production in Australia: megalitres. Includes ale and stout. Does not include beverages with alcoh"| __truncated__
#>   ..$ : chr "Quarterly beer production in Australia: megalitres. March 1956 – June 1994"
#>   .. [list output truncated]
#>  $ frequency  : num  4 12 12 12 12 12 4 12 12 12 ...
#>  $ start      : num  1948 1972 1971 1960 1967 ...
#>  $ subject    : chr  "Macro-Economic" "Industry" "Utilities" "Sales" ...