tsdl {tsdl}	R Documentation

Time Series Data Library data

Description

The time series from the Time Series Data Library.

Usage

tsdl

Format

tsdl is a list of 648 series of class tsdl. Each series within tsdl is of class ts with the following attributes:

tsp

Start time in time units, End time and Frequency. See ?tsp for details.

class

Class of ts

source

Source of the time series.

description

A short description of the time series.

subject

Subject (or type) of the time series.

Author(s)

Rob Hyndman

Yangzhuoran Yang

Source

https://datamarket.com/data/list/?q=provider:tsdl.

References

Hyndman, R.J. "Time Series Data Library", https://datamarket.com/data/list/?q=provider:tsdl.

See Also

meta_tsdl

Examples

tsdl
subset(tsdl,"Sales")

[Package tsdl version 0.1.0 Index]